requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file development.txt requirements.in development.in
#
coverage==4.4.1
decorator==4.0.11         # via ipython, traitlets
django-bootstrap3==8.2.3
django-crontab==0.7.1
django-debug-toolbar==1.8
django-registration==2.2
django-simple-captcha==0.5.5
django==1.11.1
ipython-genutils==0.2.0   # via traitlets
ipython==6.1.0
jedi==0.10.2              # via ipython
olefile==0.44             # via pillow
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
pillow==4.1.1             # via django-simple-captcha
prompt-toolkit==1.0.14    # via ipython
psycopg2==2.7.1
ptyprocess==0.5.1         # via pexpect
pygments==2.2.0           # via ipython
pytz==2017.2              # via django
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-simple-captcha, prompt-toolkit, traitlets
sqlparse==0.2.3           # via django-debug-toolbar
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
